makespan debug

+++ src/main/java/entity/Type.java
@@ -5,7 +5,7 @@ public class Type {
     public double cu;
     public double bw;
     public double p;
-    public Type(int i, float cu, float bw, float p){
+    public Type(int i, double cu, double bw, double p){
         this.id = i;
         this.cu = cu;
         this.bw = bw;
